+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.5.35 Release:
+## Bug Fixes:
+- Maneuver: Distracting Strike
+- Dread Ambusher
+
 # 0.5.34 Release:
 ## Bug Fixes:
 - Sanctuary